fix: extract secondary functions to utility class
