+++ src/components/Network/NetworksList.tsx
@@ -1,11 +1,11 @@
 import { Text } from "@radix-ui/themes"
 import clsx from "clsx"
 import type { ReactNode } from "react"
-import { getBlockchainsOptions } from "src/features/deposit/components/DepositForm"
 import type { BlockchainEnum } from "src/sdk/poaBridge/constants/blockchains"
 import type { SupportedChainName } from "src/types/base"
 import { reverseAssetNetworkAdapter } from "src/utils/adapters"
 import { isAuroraVirtualChain } from "src/utils/blockchain"
+import { getBlockchainsOptions } from "src/utils/blockchainOptions"
 import { PoweredByAuroraLabel } from "../PoweredByAuroraLabel"
 
 interface NetworkListProps {
